chore(ai): split out the tools into their own files
Also collocate the test for each tool in the same file.
This should make it easier to find and work on each tool individually
